+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/trixnity/messenger/compose/view/settings/SettingsAccountCard.kt
@@ -9,23 +9,41 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.ElevatedCard
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import de.connect2x.trixnity.core.model.UserId
+import de.connect2x.trixnity.messenger.compose.view.DI
 import de.connect2x.trixnity.messenger.compose.view.common.modifier.focusHighlighting
+import de.connect2x.trixnity.messenger.compose.view.get
 import de.connect2x.trixnity.messenger.compose.view.theme.components
 import de.connect2x.trixnity.messenger.compose.view.theme.components.ThemedSelectableText
+import de.connect2x.trixnity.messenger.compose.view.util.UserIdText
+
+interface SettingsAccountCard {
+    @Composable fun create(userId: UserId, modifier: Modifier, content: @Composable ColumnScope.() -> Unit)
+}
 
 @Composable
 fun SettingsAccountCard(userId: UserId, modifier: Modifier = Modifier, content: @Composable ColumnScope.() -> Unit) {
-    ElevatedCard(modifier.padding(bottom = 10.dp)) {
-        Column(Modifier.padding(10.dp)) {
-            Text(userId.full, style = MaterialTheme.typography.titleMedium, modifier = Modifier.focusHighlighting())
-            HorizontalDivider(Modifier.fillMaxWidth().padding(vertical = 10.dp))
-            content()
+    DI.get<SettingsAccountCard>().create(userId, modifier, content)
+}
+
+class SettingsAccountCardImpl : SettingsAccountCard {
+    @Composable
+    override fun create(userId: UserId, modifier: Modifier, content: @Composable (ColumnScope.() -> Unit)) {
+        ElevatedCard(modifier.padding(bottom = 10.dp)) {
+            Column(Modifier.padding(10.dp)) {
+                UserIdText(
+                    userId = userId,
+                    showTooltip = false,
+                    style = MaterialTheme.typography.titleMedium,
+                    modifier = Modifier.focusHighlighting(),
+                )
+                HorizontalDivider(Modifier.fillMaxWidth().padding(vertical = 10.dp))
+                content()
+            }
         }
     }
 }