+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/trixnity/messenger/compose/view/util/UserIdText.kt
@@ -1,0 +1,199 @@
+package de.connect2x.trixnity.messenger.compose.view.util
+
+import androidx.compose.foundation.text.TextAutoSize
+import androidx.compose.material3.LocalTextStyle
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.semantics.semantics
+import androidx.compose.ui.text.TextLayoutResult
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontStyle
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.text.style.TextDecoration
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.TextUnit
+import de.connect2x.trixnity.core.model.UserId
+import de.connect2x.trixnity.messenger.compose.view.DI
+import de.connect2x.trixnity.messenger.compose.view.common.Tooltip
+import de.connect2x.trixnity.messenger.compose.view.get
+import de.connect2x.trixnity.messenger.compose.view.theme.components.SelectionStyle
+import de.connect2x.trixnity.messenger.compose.view.theme.components.ThemedSelectionContainer
+
+interface UserIdText {
+    @Composable
+    fun create(
+        userId: UserId,
+        showTooltip: Boolean,
+        modifier: Modifier,
+        color: Color,
+        autoSize: TextAutoSize?,
+        fontSize: TextUnit,
+        fontStyle: FontStyle?,
+        fontWeight: FontWeight?,
+        fontFamily: FontFamily?,
+        letterSpacing: TextUnit,
+        textDecoration: TextDecoration?,
+        textAlign: TextAlign?,
+        lineHeight: TextUnit,
+        overflow: TextOverflow,
+        softWrap: Boolean,
+        maxLines: Int,
+        minLines: Int,
+        onTextLayout: ((TextLayoutResult) -> Unit)?,
+        style: TextStyle,
+    )
+}
+
+@Composable
+fun UserIdText(
+    userId: UserId,
+    showTooltip: Boolean = true,
+    modifier: Modifier = Modifier,
+    color: Color = Color.Unspecified,
+    autoSize: TextAutoSize? = null,
+    fontSize: TextUnit = TextUnit.Unspecified,
+    fontStyle: FontStyle? = null,
+    fontWeight: FontWeight? = null,
+    fontFamily: FontFamily? = null,
+    letterSpacing: TextUnit = TextUnit.Unspecified,
+    textDecoration: TextDecoration? = null,
+    textAlign: TextAlign? = null,
+    lineHeight: TextUnit = TextUnit.Unspecified,
+    overflow: TextOverflow = TextOverflow.Clip,
+    softWrap: Boolean = true,
+    maxLines: Int = Int.MAX_VALUE,
+    minLines: Int = 1,
+    onTextLayout: ((TextLayoutResult) -> Unit)? = null,
+    style: TextStyle = LocalTextStyle.current,
+) {
+    DI.get<UserIdText>()
+        .create(
+            userId = userId,
+            showTooltip = showTooltip,
+            modifier = modifier,
+            color = color,
+            autoSize = autoSize,
+            fontSize = fontSize,
+            fontStyle = fontStyle,
+            fontWeight = fontWeight,
+            fontFamily = fontFamily,
+            letterSpacing = letterSpacing,
+            textDecoration = textDecoration,
+            textAlign = textAlign,
+            lineHeight = lineHeight,
+            overflow = overflow,
+            softWrap = softWrap,
+            maxLines = maxLines,
+            minLines = minLines,
+            onTextLayout = onTextLayout,
+            style = style,
+        )
+}
+
+@Composable
+fun ThemedSelectableUserIdText(
+    userId: UserId,
+    showTooltip: Boolean = true,
+    selectionStyle: SelectionStyle,
+    selectionModifier: Modifier = Modifier,
+    modifier: Modifier = Modifier,
+    color: Color = Color.Unspecified,
+    autoSize: TextAutoSize? = null,
+    fontSize: TextUnit = TextUnit.Unspecified,
+    fontStyle: FontStyle? = null,
+    fontWeight: FontWeight? = null,
+    fontFamily: FontFamily? = null,
+    letterSpacing: TextUnit = TextUnit.Unspecified,
+    textDecoration: TextDecoration? = null,
+    textAlign: TextAlign? = null,
+    lineHeight: TextUnit = TextUnit.Unspecified,
+    overflow: TextOverflow = TextOverflow.Clip,
+    softWrap: Boolean = true,
+    maxLines: Int = Int.MAX_VALUE,
+    minLines: Int = 1,
+    onTextLayout: ((TextLayoutResult) -> Unit)? = null,
+    style: TextStyle = LocalTextStyle.current,
+) {
+    ThemedSelectionContainer(selectionStyle, selectionModifier.semantics(mergeDescendants = true) {}) {
+        UserIdText(
+            userId = userId,
+            showTooltip = showTooltip,
+            modifier = modifier,
+            color = color,
+            autoSize = autoSize,
+            fontSize = fontSize,
+            fontStyle = fontStyle,
+            fontWeight = fontWeight,
+            fontFamily = fontFamily,
+            letterSpacing = letterSpacing,
+            textDecoration = textDecoration,
+            textAlign = textAlign,
+            lineHeight = lineHeight,
+            overflow = overflow,
+            softWrap = softWrap,
+            maxLines = maxLines,
+            minLines = minLines,
+            onTextLayout = onTextLayout,
+            style = style,
+        )
+    }
+}
+
+class UserIdTextImpl : UserIdText {
+    @Composable
+    override fun create(
+        userId: UserId,
+        showTooltip: Boolean,
+        modifier: Modifier,
+        color: Color,
+        autoSize: TextAutoSize?,
+        fontSize: TextUnit,
+        fontStyle: FontStyle?,
+        fontWeight: FontWeight?,
+        fontFamily: FontFamily?,
+        letterSpacing: TextUnit,
+        textDecoration: TextDecoration?,
+        textAlign: TextAlign?,
+        lineHeight: TextUnit,
+        overflow: TextOverflow,
+        softWrap: Boolean,
+        maxLines: Int,
+        minLines: Int,
+        onTextLayout: ((TextLayoutResult) -> Unit)?,
+        style: TextStyle,
+    ) {
+        @Composable
+        fun UserIdTextInner() {
+            Text(
+                text = userId.full,
+                modifier = modifier,
+                color = color,
+                autoSize = autoSize,
+                fontSize = fontSize,
+                fontStyle = fontStyle,
+                fontWeight = fontWeight,
+                fontFamily = fontFamily,
+                letterSpacing = letterSpacing,
+                textDecoration = textDecoration,
+                textAlign = textAlign,
+                lineHeight = lineHeight,
+                overflow = overflow,
+                softWrap = softWrap,
+                maxLines = maxLines,
+                minLines = minLines,
+                onTextLayout = onTextLayout,
+                style = style,
+            )
+        }
+
+        if (showTooltip) {
+            Tooltip({ Text(userId.full) }) { UserIdTextInner() }
+        } else {
+            UserIdTextInner()
+        }
+    }
+}