+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/DI.kt
@@ -29,6 +29,8 @@ import de.connect2x.trixnity.messenger.util.DragAndDropHandler
 import de.connect2x.trixnity.messenger.util.DragAndDropHandlerBase
 import de.connect2x.trixnity.messenger.util.EnterRoom
 import de.connect2x.trixnity.messenger.util.EnterRoomImpl
+import de.connect2x.trixnity.messenger.util.GetAccountProfileDisplayName
+import de.connect2x.trixnity.messenger.util.GetAccountProfileDisplayNameImpl
 import de.connect2x.trixnity.messenger.util.InformationMarkdownFlavour
 import de.connect2x.trixnity.messenger.util.InformationMarkdownFlavourImpl
 import de.connect2x.trixnity.messenger.util.LeaveRoom
@@ -296,6 +298,7 @@ fun createTrixnityMessengerDefaultModuleFactories(): List<ModuleFactory> =
                 single<SharedDataHandler> { SharedDataHandlerImpl() }
 
                 single<BackHandler> { BackHandlerImpl() }
+                single<GetAccountProfileDisplayName> { GetAccountProfileDisplayNameImpl(get()) }
             }
         },
         ::connectingViewModels,