+++ CHANGELOG.md
@@ -53,6 +53,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 - CI, INTERNAL: Enabled more detekt linting rules
 - SDK,UI: Add better error handling with UI feedback to the `AudioRecorder` API
 - SDK,UI: Experimental androidx.navigation3 support
+- SDK: Add `GetAccountProfileDisplayName` and `rememberAccountProfileDisplayName` to acquire the username from a logged-in `MatrixClient`
+- SDK,UI: Add `CopyableUserId` providing a selectable text especially for user MXIDs over the DI
+- SDK,UI: Add `UserIdText` providing text components especially for user MXIDs over the DI
 - SDK,UI: Add `TwoPaneScenePlaceholder` displayed in the second pane when no room is opened
 
 ### Changed